Move other Statement in astroid.nodes.statement.py
