+++ tests/unittest_python3.py
@@ -20,8 +20,7 @@ from textwrap import dedent
 
 from astroid import nodes
 from astroid.builder import AstroidBuilder, extract_node
-from astroid.nodes.node_classes import Assign, Const, Expr, Name, YieldFrom
-from astroid.nodes.scoped_nodes import ClassDef, FunctionDef
+from astroid.nodes import Assign, ClassDef, Const, Expr, FunctionDef, Name, YieldFrom
 from astroid.test_utils import require_version
